LayoutTests/editing/pasteboard/paste-blockquote-into-blockquote-2.html

<!DOCTYPE html>
<html>
<head>
<style>
blockquote {
    color: blue;
    border-left: 2px solid blue;
    margin: 0px;
    padding: 0 0 0 20px;
}
</style>
</head>

<body>
<div contenteditable="true" id="editable"><blockquote type='cite' id="block">One</blockquote></div>

<script src="../../resources/dump-as-markup.js"></script>
<script>
var sel = window.getSelection();
var block = document.getElementById("block");

sel.setPosition(block, 1);
document.execCommand("InsertHTML", false, "<span><blockquote type='cite'><blockquote type='cite'><div>Two</div><div>Three</div></blockquote></blockquote></span>");
Markup.description("This tests pasting a nested blockquote into a blockquote. "
    + "The text \"One\" should be singly-blockquoted, and the text \"Two Three\" should be doubly-blockquoted. "
    + "However, currently \"One Two\" is singly-blockquoted and \"Three\" is doubly-blockquoted. See bug 61807.");
Markup.dump(editable);
</script>
</body>
</html>
