third_party/blink/web_tests/platform/generic/editing/pasteboard/paste-blockquote-into-blockquote-2-expected.txt

This tests pasting a nested blockquote into a blockquote. The text "One" should be singly-blockquoted, and the text "Two Three" should be doubly-blockquoted. However, currently "One Two" is singly-blockquoted and "Three" is doubly-blockquoted. See bug 61807.
| <blockquote>
|   id="block"
|   type="cite"
|   "OneTwo"
| <blockquote>
|   type="cite"
|   <blockquote>
|     type="cite"
|     <div>
|       "Three<#selection-caret>"
